requirements-dev.txt
# Development requirements for the backend (Tagline)
-r requirements.txt

# Testing
pytest>=8.1.0,<9.0.0
pytest-cov>=5.0.0,<6.0.0
pytest-order>=1.2.0,<2.0.0
requests>=2.31.0,<3.0.0
requests-mock
httpx>=0.27.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0  # Load .env files

# Linting/Formatting
isort>=6.0.1,<7.0.0
black>=24.3.0,<25.0.0
ruff>=0.3.0,<1.0.0

# Type checking
pyright==1.1.400

# Pre-commit
pre-commit>=3.7.0,<4.0.0

# Docker/devops tools for E2E tests
docker>=7.0.0,<8.0.0

# Task queue (dev)
rq>=1.16.1
rq-dashboard>=0.8.2.2
